requirements.txt
aiofiles
dnspython
